refactor(core): split geometry builders and thin runPipeline phases
Break the ~1k-line geometry module into shared primitives, simple marks,
and composite builders, keeping a thin dispatch facade. Also extract
preparePanels and trainPipelineScales so runPipeline is a short phase
chain. Characterization tests cover geom dispatch and panel preparation.
